% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT5G62640.R
\docType{data}
\name{AT5G62640}
\alias{AT5G62640}
\title{Gene model for AT5G62640}
\format{A data frame with 32 rows and 2 variables:
\describe{
  \item{type}{feature of gene ie intron exon or UTR}
  \item{coordinates}{start and stop of gene feature separated by a "-"}
  ...
}}
\source{
\url{https://www.arabidopsis.org/servlets/TairObject?type=gene&id=1000654517}
}
\usage{
AT5G62640
}
\description{
Gene model for AT5G62640
}
\keyword{datasets}

