\name{genemodel-package}
\alias{genemodel-package}
\alias{genemodel}
\docType{package}
\title{
\packageTitle{genemodel}
}
\description{
\packageDescription{genemodel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{genemodel}
\packageIndices{genemodel}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{genemodel}

Maintainer: \packageMaintainer{genemodel}
}
\references{
 Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simple examples of the most important functions
}
