% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemodel.plot.R
\name{genemodel.plot}
\alias{genemodel.plot}
\title{genemodel.plot}
\usage{
genemodel.plot(model, start, bpstop, orientation)
}
\arguments{
\item{model}{data.frame containing model information. Required columns are "feature", "start"}

\item{start}{start position}

\item{bpstop}{stop position}

\item{orientation}{either "foward" or "reverse" indicates the direction of transcription}
}
\description{
This function plots a gene model
}
\examples{


}

