% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print hello}
\usage{
hello(fname, lname)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function prints hello
}
\examples{
hello(fname="Tomas",lname="Greif")
hello(fname="Your",lname="Name")
}

