% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemodel.plot.R
\name{genemodel.plot}
\alias{genemodel.plot}
\title{genemodel.plot}
\usage{
genemodel.plot(model, start, bpstop, orientation, xaxis = TRUE)
}
\arguments{
\item{model}{data.frame containing model information. Required columns are "type", "coordinates"}

\item{start}{start position}

\item{bpstop}{stop position}

\item{orientation}{either "foward" or "reverse" indicates the direction of transcription}

\item{xaxis}{default is TRUE and adds axis above gene model showing position}
}
\description{
This function plots a gene model
}
\examples{
data(AT5G62640)
genemodel.plot(AT5G62640, 25149433, 25152541, "reverse", xaxis=TRUE)
}

