% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.plot.R
\name{mutation.plot}
\alias{mutation.plot}
\title{mutation.plot}
\usage{
mutation.plot(start, stop, text = "", drop = -0.15, col = "red",
  haplotypes = NULL)
}
\arguments{
\item{start}{start position}

\item{stop}{stop position}

\item{text}{any text that you want displayed on the label}

\item{drop}{how far below the gene model you want the mutation label to be placed}

\item{col}{the color of the text and mutation line to be}

\item{haplotypes}{the color of dots that you want to place along the mutation line to indicate some factor such as haplotype that the mutation belongs to}
}
\description{
This function plots mutations along genemodels created with genemodel.plot
}
\examples{
data(AT5G62640)
genemodel.plot(AT5G62640, 25149433, 25152541, "reverse")
mutation.plot(25149593, 25149593, text="P->S", col="red", haplotype="blue")
}

